Configurable Controls
Added the ability to configure controls to any key. To see the demo, go
to the New Game page and try pressing the control buttons. The canvas
changes to a different color for each button. Then go to the Controls
page and change some controls and go back to the New Game page and try
it out again!

I also added some of the gameLoop structure. I needed to do that in
order to update the canvas when the keys were pressed.